move pogo classes to data access library
