+++ DataAccessLibrary/DataAccessLibrary.csproj
@@ -16,8 +16,4 @@
     </Reference>
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="DataModels\" />
-  </ItemGroup>
-
 </Project>
