+++ StudySmarterFlashcards/ImportTools/ImportFlashcardService.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
-using StudySmarterFlashcards.Sets;
+using DataAccessLibrary.DataModels;
 using Syncfusion.DocIO.DLS;
 using Syncfusion.XlsIO;
 using Windows.Storage;
