+++ StudySmarterFlashcards/Sets/SetViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using DataAccessLibrary.DataModels;
 using GalaSoft.MvvmLight.Command;
 using GalaSoft.MvvmLight.Messaging;
 using GalaSoft.MvvmLight.Views;
