+++ StudySmarterFlashcards/Study/FillBlankStudyViewModel.cs
@@ -5,17 +5,15 @@ using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Windows.Input;
+using DataAccessLibrary.DataModels;
 using GalaSoft.MvvmLight.Command;
 using GalaSoft.MvvmLight.Messaging;
 using GalaSoft.MvvmLight.Views;
 using StudySmarterFlashcards.Dialogs;
-using StudySmarterFlashcards.Sets;
 using StudySmarterFlashcards.Utils;
 using Windows.System;
 using Windows.UI.Core;
 using Windows.UI.Xaml;
-using Windows.UI.Xaml.Input;
 
 namespace StudySmarterFlashcards.Study
 {
