+++ StudySmarterFlashcards/StudySmarterFlashcards.csproj
@@ -154,12 +154,10 @@
     </Compile>
     <Compile Include="Menus\SettingsViewModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Sets\CardSetModel.cs" />
     <Compile Include="Sets\EditSetPage.xaml.cs">
       <DependentUpon>EditSetPage.xaml</DependentUpon>
     </Compile>
     <Compile Include="Sets\EditSetViewModel.cs" />
-    <Compile Include="Sets\IndividualCardModel.cs" />
     <Compile Include="Sets\SetPage.xaml.cs">
       <DependentUpon>SetPage.xaml</DependentUpon>
     </Compile>
@@ -232,6 +230,9 @@
     <Content Include="Assets\Wide310x150Logo.scale-150.png" />
     <Content Include="Assets\Wide310x150Logo.scale-200.png" />
     <Content Include="Assets\Wide310x150Logo.scale-400.png" />
+    <None Include="..\.editorconfig">
+      <Link>.editorconfig</Link>
+    </None>
     <None Include="Package.StoreAssociation.xml" />
     <Content Include="Properties\Default.rd.xml" />
   </ItemGroup>
